+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+1.5.10
+============
+* online form improvements && fixes
+* package searchDelay
+* cashDocuments - search by payer
+
+1.5.9
+=============
+* new rabbitmq (redmine #644)
+* fix translations
+
 1.5.8
 =============
 * booking.com: fix (redmine #717)